Remove main.rb. Update config files and alfeal lib
