+++ Gemfile
@@ -3,6 +3,6 @@ source 'https://rubygems.org'
 gem 'sinatra'
 
 group :devel do
-  gem 'pry'
+  gem 'byebug'
   gem 'rubocop'
 end
