+++ config/config.yml
@@ -1,54 +1,61 @@
-conjugation:
-  -
-tense:
-  - s: الماضي
-    q: p
-  - s: المضارع المرفوع
-    q: F
-  - s: المضارع المنصوب
-    q: m
-    tag: mansoub
-  - s: المضارع المجزوم
-    q: m
-    tag: majssum
-  - s: الأمر
-    q: i
-number:
-  - المفرد
-  - المثنّى
-  - الجمع
-person:
-  - المتكلّم
-  - المخاطب
-  - المخاطبة
-  - الغائب
-  - الغائبة
-verb:
-  - s: دهب يذهب
-    q: دَهَبَ-فتحة-متعدي
-  - s: خَرَجَ يخرُجُ
-    q: خَرَجَ-ضمة-متعدي
-  - s: رجعَ يرجِعُ
-    q: رَجَعَ-كسرة-متعدي
-  - s: فَهِمَ يَفْهَمُ
-    q: فَهِمَ-فتحة-متعدي
-  - s: سافر يُسافِرُ
-    q: سَفَرَ-كسرة-متعدي
-  - s:  أحَبَّ يُحِبُّ
-    q: أحَبَّ-كسرة-متعدي
-  - s:  دَرَسَ يَدْرُسُ
-    q: دَرَسَ-ضمة-متعدي
-  - s: سكن يسكن
-    q: سَكَنَ-ضمة-متعدي
-  - s:  زَارَ يَزورُ
-    q: زَارَ-ضمة-متعدي
-  - s:   قال يَقولُ
-    q: قَالَ-ضمة-متعدي
-  - s:  كان يَكون
-    q: كَانَ-ضمة-متعدي
-  - s:  نامَ يَنامُ
-    q: نَامَ-فتحة-متعدي
-  - s:  خاف يَخاف
-    q: خَافَ-فتحة-متعدي
-  - s: عاشَ يَعيَش
-    q: عَاشَ-كسرة-متعدي
+# encoding: utf-8
+
+default: &default_settings
+  conjugation:
+    -
+  tense:
+    - s: الماضي
+      q: p
+    - s: المضارع المرفوع
+      q: F
+    - s: المضارع المنصوب
+      q: m
+      tag: mansoub
+    - s: المضارع المجزوم
+      q: m
+      tag: majssum
+    - s: الأمر
+      q: i
+  number:
+    - المفرد
+    - المثنّى
+    - الجمع
+  person:
+    - المتكلّم
+    - المخاطب
+    - المخاطبة
+    - الغائب
+    - الغائبة
+  verb:
+    - s: دهب يذهب
+      q: دَهَبَ-فتحة-متعدي
+    - s: خَرَجَ يخرُجُ
+      q: خَرَجَ-ضمة-متعدي
+    - s: رجعَ يرجِعُ
+      q: رَجَعَ-كسرة-متعدي
+    - s: فَهِمَ يَفْهَمُ
+      q: فَهِمَ-فتحة-متعدي
+    - s: سافر يُسافِرُ
+      q: سَفَرَ-كسرة-متعدي
+    - s:  أحَبَّ يُحِبُّ
+      q: أحَبَّ-كسرة-متعدي
+    - s:  دَرَسَ يَدْرُسُ
+      q: دَرَسَ-ضمة-متعدي
+    - s: سكن يسكن
+      q: سَكَنَ-ضمة-متعدي
+    - s:  زَارَ يَزورُ
+      q: زَارَ-ضمة-متعدي
+    - s:   قال يَقولُ
+      q: قَالَ-ضمة-متعدي
+    - s:  كان يَكون
+      q: كَانَ-ضمة-متعدي
+    - s:  نامَ يَنامُ
+      q: نَامَ-فتحة-متعدي
+    - s:  خاف يَخاف
+      q: خَافَ-فتحة-متعدي
+    - s: عاشَ يَعيَش
+      q: عَاشَ-كسرة-متعدي
+
+development:
+  <<: *default_settings
+
